+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + Improved Arcane workbench category
 + Improved Infusion category
 + Improved Aspect compound category
++ Improved Aspect from itemstack category
 
 # 1.3.1
 + Added a check for items that dont have aspects to dont render in the tooltip